Modify tests 12 and 14 to match new shadow init
Test 12 checks the Horse's starting shadows. After the new shadow
implementation, the order for some of the expected shadows was
incorrect, so I ordered them according to Horse's update_shadows
method.

Test 14 checks the Cannon's starting shadows. The test originally did
not account for the mistake in shadow initialization I had made with
the former shadow implementation, so I added the new expected shadows
for the east and west points.