Refactor repetitive code for river crossing
Previously, I used an if-else structure to check if the soldier had
or hadn't crossed the river, and this created repetitive code for
checking the south and north points of the red and black soldiers,
respectively.

The modified code only checks if the river has been crossed, and if
so, checks the soldier's left and right points. The else statement
has been removed because by default the soldier should be able to
move forward regardless of whether it has crossed the river or not.